requirements.txt
numpy
scipy
matplotlib
tqdm
scikit-learn
scikit-image
pillow
SimpleITK
tensorflow-gpu
tensorflow.keras
h5py